@@ -11,48 +11,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.14",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
+    "@angular/animations": "~12.1.0",
+    "@angular/cdk": "^12.1.1",
+    "@angular/common": "~12.1.0",
+    "@angular/compiler": "~12.1.0",
+    "@angular/core": "~12.1.0",
+    "@angular/flex-layout": "^12.0.0-beta.34",
+    "@angular/forms": "~12.1.0",
+    "@angular/material": "^12.1.1",
+    "@angular/platform-browser": "~12.1.0",
+    "@angular/platform-browser-dynamic": "~12.1.0",
+    "@angular/router": "~12.1.0",
+    "@ng-bootstrap/ng-bootstrap": "^10.0.0",
     "angular-popper": "^2.0.1",
-    "bootstrap": "4.5.0",
+    "bootstrap": "^4.5.0",
     "core-js": "^3.6.5",
     "jquery": "^3.5.1",
     "jquery-sparkline": "^2.4.0",
-    "ng2-slim-loading-bar": "^4.0.0",
-    "ngx-spinner": "^9.0.2",
+    "moment": "^2.29.1",
+    "ngx-spinner": "^12.0.0",
     "ngx-toastr": "^12.1.0",
     "popper.js": "^1.16.1",
-    "rxjs": "~6.4.0",
+    "rxjs": "~6.5.3",
     "rxjs-compat": "^6.3.3",
     "sweetalert2": "^9.15.3",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "tslib": "^2.2.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.22",
-    "@angular/cli": "~8.3.22",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~3.3.8",
+    "@angular-devkit/build-angular": "~12.1.0",
+    "@angular/cli": "~12.1.0",
+    "@angular/compiler-cli": "~12.1.0",
+    "@angular/language-service": "~12.1.0",
+    "@types/node": "^12.11.1",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
+    "karma": "~6.3.0",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~5.4.4",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "typescript": "~4.3.2"
   }
 }
