@@ -1,13 +1,17 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import {EmployeeComponent} from './components/employee/employee.component';
-import {FormEmployeeComponent} from './components/form-employee/form-employee.component';
+import {FormEmployeeComponent} from './components/employee/form-employee/form-employee.component';
 import {HomeComponent} from './components/home/home.component';
 
 const routes: Routes = [
   {path: 'home', component: HomeComponent},
-  {path: 'employee', component: EmployeeComponent},
-  {path: 'form-employee/:id', component: FormEmployeeComponent},
+  {
+    path: 'employee', component: EmployeeComponent, children: [
+      { path: 'new', component: FormEmployeeComponent },
+      { path: 'edit/:id', component: FormEmployeeComponent }
+    ]
+  }
 ];
 
 @NgModule({
