@@ -6,7 +6,7 @@ public interface CRUD<T> {
 
 	List<T> listAll();
 	T findById(Long id);
-	void save(T object);
+	T save(T object);
 	void update(T object);
 	void delete(Long id);
 }
